+++ packages/jetbrains-plugin/src/main/kotlin/com/mongodb/jbplugin/editor/MongoDbVirtualFileDataSourceProvider.kt
@@ -8,30 +8,45 @@ import com.intellij.database.util.VirtualFileDataSourceProvider
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Key
 import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
+import com.mongodb.jbplugin.dialects.Dialect
+
+private const val KEY_PREFIX = "com.mongodb.jbplugin"
 
 val PsiFile.dataSource: LocalDataSource?
-    get() =
+    get() = runCatching {
         MongoDbVirtualFileDataSourceProvider()
             .getDataSource(
                 project,
                 virtualFile,
             )?.localDataSource
+    }.getOrNull()
 
 val PsiFile.database: String?
-    get() =
+    get() = runCatching {
         MongoDbVirtualFileDataSourceProvider()
             .getDatabase(
                 virtualFile,
             )
+    }.getOrNull()
+
+val PsiFile.dialect: Dialect<PsiElement, Project>?
+    get() = runCatching {
+        MongoDbVirtualFileDataSourceProvider()
+            .getDialect(
+                virtualFile,
+            )
+    }.getOrNull()
 
 /**
  * Returns the data source, if attached to the editor through the MongoDB Plugin.
  */
 class MongoDbVirtualFileDataSourceProvider : VirtualFileDataSourceProvider() {
     object Keys {
-        internal val attachedDataSource: Key<LocalDataSource> = Key.create("com.mongodb.jbplugin.AttachedDataSource")
-        internal val attachedDatabase: Key<String> = Key.create("com.mongodb.jbplugin.AttachedDatabase")
+        internal val attachedDataSource: Key<LocalDataSource> = Key.create("$KEY_PREFIX.AttachedDataSource")
+        internal val attachedDatabase: Key<String> = Key.create("$KEY_PREFIX.AttachedDatabase")
+        internal val attachedDialect: Key<Dialect<PsiElement, Project>> = Key.create("$KEY_PREFIX.AttachedDialect")
     }
 
     override fun getDataSource(
@@ -47,4 +62,8 @@ class MongoDbVirtualFileDataSourceProvider : VirtualFileDataSourceProvider() {
     fun getDatabase(
         file: VirtualFile
     ): String? = file.getUserData(Keys.attachedDatabase)
+
+    fun getDialect(
+        file: VirtualFile
+    ): Dialect<PsiElement, Project>? = file.getUserData(Keys.attachedDialect)
 }