+++ packages/jetbrains-plugin/src/main/kotlin/com/mongodb/jbplugin/inspections/AbstractMongoDbInspectionBridge.kt
@@ -5,29 +5,42 @@ import com.intellij.codeInspection.LocalInspectionToolSession
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.database.dataSource.localDataSource
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Key
-import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.*
 import com.intellij.psi.util.CachedValue
 import com.intellij.psi.util.CachedValueProvider
 import com.intellij.psi.util.CachedValuesManager
 import com.intellij.psi.util.PsiTreeUtil
 import com.mongodb.jbplugin.dialects.Dialect
-import com.mongodb.jbplugin.editor.MongoDbVirtualFileDataSourceProvider
+import com.mongodb.jbplugin.editor.dataSource
+import com.mongodb.jbplugin.editor.database
+import com.mongodb.jbplugin.editor.dialect
 import com.mongodb.jbplugin.mql.Node
 
 /**
- * @param dialect
+ * This class is used to connect a MongoDB inspection to IntelliJ.
+ * It's responsible for getting the dialect of the current file and
+ * do the necessary dependency injection to make the inspection work.
+ *
+ * Usually you won't reimplement methods, just create a new empty class
+ * that provides the inspection implementation, in the same file.
+ *
+ * @see com.mongodb.jbplugin.inspections.impl.FieldCheckInspectionBridge as an example
+ *
  * @param inspection
  */
+@Suppress("TOO_LONG_FUNCTION")
 abstract class AbstractMongoDbInspectionBridge(
-    private val dialect: Dialect<PsiElement>,
     private val inspection: MongoDbInspection,
 ) : AbstractBaseJavaLocalInspectionTool() {
-    private val queryKey: Key<CachedValue<Node<PsiElement>>> =
-        Key.create(
-            "QueryForDialect${dialect.javaClass.name}",
+    private val queryKeysByDialect = mutableMapOf<Dialect<PsiElement, Project>, Key<CachedValue<Node<PsiElement>>>>()
+    private fun queryKey(dialect: Dialect<PsiElement, Project>) =
+        queryKeysByDialect.getOrPut(dialect) {
+ Key.create(
+            "QueryForDialect${dialect.javaClass.name}"
         )
+}
 
     /**
      * Ktlint complains about buildVisitor being longer than 50 lines but because it is just an object implementation
@@ -50,9 +63,17 @@ abstract class AbstractMongoDbInspectionBridge(
 
             private fun dispatchIfValidMongoDbQuery(expression: PsiElement) {
                 ApplicationManager.getApplication().runReadAction {
+                    val dialect = expression.containingFile.dialect ?: return@runReadAction
+                    val queryKey = queryKey(dialect)
+
                     var cachedValue: CachedValue<Node<PsiElement>>? = null
                     if (dialect.parser.isCandidateForQuery(expression)) {
                         val attachment = dialect.parser.attachment(expression)
+                        val psiManager = PsiManager.getInstance(expression.project)
+                        if (!psiManager.areElementsEquivalent(expression, attachment)) {
+                            return@runReadAction
+                        }
+
                         attachment.getUserData(queryKey)?.let {
                             cachedValue = attachment.getUserData(queryKey)!!
                         } ?: run {
@@ -72,16 +93,12 @@ abstract class AbstractMongoDbInspectionBridge(
                             inspection.visitMongoDbQuery(null, holder, cachedValue!!.value, dialect.formatter)
                         } else {
                             val cachedQuery = cachedValue!!.value
-                            val dataSource =
-                                MongoDbVirtualFileDataSourceProvider().getDataSource(
-                                    expression.project,
-                                    fileInExpression.virtualFile,
-                                )
+                            val dataSource = fileInExpression.dataSource
 
                             inspection.visitMongoDbQuery(
                                 dataSource?.localDataSource,
                                 holder,
-                                queryWithCollectionReference(cachedQuery, fileInExpression.virtualFile),
+                                queryWithCollectionReference(cachedQuery, fileInExpression),
                                 dialect.formatter,
                             )
                         }
@@ -89,23 +106,7 @@ abstract class AbstractMongoDbInspectionBridge(
                 }
             }
 
-            private fun queryWithCollectionReference(query: Node<PsiElement>, virtualFile: VirtualFile) =
-                MongoDbVirtualFileDataSourceProvider().getDatabase(
-                    virtualFile,
-                )?.let { query.queryWithOverwrittenDatabase(it) } ?: query
+            private fun queryWithCollectionReference(query: Node<PsiElement>, psiFile: PsiFile) =
+                psiFile.database?.let { query.queryWithOverwrittenDatabase(it) } ?: query
         }
-}
-
-/**
- * Checks whether a provided problem description has already been registered with the ProblemsHolder for a given
- * PsiElement
- * Warning: Instead of using this, we should get around fixing the "possible" underlying issue highlighted by
- * INTELLIJ-60
- *
- * @param problem - Description of the problem
- * @param source - PsiElement for which the problem is to be checked
- * @return Boolean
- */
-fun ProblemsHolder.isProblemAlreadyRegistered(problem: String, source: PsiElement): Boolean = this.results.any {
-    it.psiElement == source && it.descriptionTemplate == problem
 }